services/event-collector/requirements.txt
# Runtime dependencies
redis[hiredis]>=5.0
asyncpg>=0.29
pydantic>=2.5
nats-py>=2.7.0

# Dev/test dependencies
pytest>=8.0
pytest-asyncio>=0.23
